Add functionality to save and export simulations
