+++ src/main/java/com/project/demo/rest/ivaSimulationExport/IvaSimulationExport.java
@@ -1,59 +1,58 @@
-package com.project.demo.rest.SimulationExport;
+package com.project.demo.rest.ivaSimulationExport;
 
 import com.itextpdf.html2pdf.ConverterProperties;
 import com.itextpdf.html2pdf.HtmlConverter;
 import com.itextpdf.kernel.pdf.PdfDocument;
 import com.itextpdf.kernel.pdf.PdfWriter;
-import com.project.demo.logic.entity.isrSimulation.IsrRepository;
-import com.project.demo.logic.entity.isrSimulation.IsrSimulation;
-import com.project.demo.logic.entity.isrSimulation.SimulationExportService;
+import com.project.demo.logic.entity.ivacalculation.IvaCalculation;
+import com.project.demo.logic.entity.ivacalculation.IvaCalculationRepository;
+import com.project.demo.logic.entity.ivacalculation.IvaExportService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.*;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
 import java.io.ByteArrayOutputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.Optional;
 
-@RestController
-@RequestMapping("/export")
-public class SimulationExportController {
+@Controller
+@RequestMapping("/iva-export")
+public class IvaSimulationExport {
 
     @Autowired
-    IsrRepository isrRepository;
+    IvaCalculationRepository ivaCalculationRepository;
 
     @Autowired
-    SimulationExportService simulationExportService;
+    IvaExportService ivaExportService;
 
     @GetMapping("/generate-pdf/{id}")
     @PreAuthorize("hasAnyRole('SUPER_ADMIN', 'USER')")
     public ResponseEntity<byte[]> generatePdf(@PathVariable Long id) {
-        Optional<IsrSimulation> foundSimulation = isrRepository.findById(id);
+        Optional<IvaCalculation> foundSimulation = ivaCalculationRepository.findById(id);
 
         if (foundSimulation.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        IsrSimulation simulation = foundSimulation.get();
+
+        IvaCalculation simulation = foundSimulation.get();
 
         try {
-            String htmlBody = simulationExportService.generateSimulationPdf(simulation);
-            String formattedHtmlContent = simulationExportService.generateHtmlContent(htmlBody);
+            String htmlContent = ivaExportService.generateSimulationPdf(simulation);
 
             ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-            PdfWriter pdfWriter = new PdfWriter(outputStream);
-            PdfDocument pdfDocument = new PdfDocument(pdfWriter);
-            ConverterProperties converterProperties = new ConverterProperties();
-            HtmlConverter.convertToPdf(formattedHtmlContent, pdfDocument, converterProperties);
+            PdfWriter writer = new PdfWriter(outputStream);
+            PdfDocument pdf = new PdfDocument(writer);
+            ConverterProperties properties = new ConverterProperties();
+
+            HtmlConverter.convertToPdf(htmlContent, pdf, properties);
 
             HttpHeaders headers = new HttpHeaders();
-            headers.setContentDisposition(ContentDisposition.builder("attachment")
-                    .filename("simulación.pdf")
-                    .build());
             headers.setContentType(MediaType.APPLICATION_PDF);
+            headers.setContentDisposition(ContentDisposition.attachment().filename("IVA-Simulacion.pdf").build());
 
             return new ResponseEntity<>(outputStream.toByteArray(), headers, HttpStatus.OK);
 
@@ -64,24 +63,22 @@ public class SimulationExportController {
     }
 
     @GetMapping("/generate-csv/{id}")
+    @PreAuthorize("hasAnyRole('SUPER_ADMIN', 'USER')")
     public ResponseEntity<byte[]> generateCsv(@PathVariable Long id) {
-        Optional<IsrSimulation> foundSimulation = isrRepository.findById(id);
+        Optional<IvaCalculation> foundSimulation = ivaCalculationRepository.findById(id);
 
         if (foundSimulation.isEmpty()) {
-            return ResponseEntity.notFound().build();
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        IsrSimulation sim = foundSimulation.get();
-        String csvContent = simulationExportService.generateSimulationCsv(sim);
+
+        IvaCalculation simulation = foundSimulation.get();
+        String csvContent = ivaExportService.generateCsv(simulation);
         byte[] csvBytes = csvContent.getBytes(StandardCharsets.UTF_8);
 
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.parseMediaType("text/csv"));
-        headers.setContentDisposition(ContentDisposition.builder("attachment")
-                .filename("Simulación.csv")
-                .build());
+        headers.setContentDisposition(ContentDisposition.attachment().filename("IVA-Simulacion.csv").build());
 
         return new ResponseEntity<>(csvBytes, headers, HttpStatus.OK);
     }
 }
-
-
